src/requirements.txt
django==1.9.5
django-widget-tweaks==1.4.1
Pillow==3.2.0
django-sticky-uploads==0.4.0